+++ src/Api/GetOrderDetailsRequest.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace ProtoneMedia\LaravelPaddle\Api;
+
+class GetOrderDetailsRequest extends Request
+{
+    public function url(): string
+    {
+        return "https://checkout.paddle.com/api{$this->uri}";
+    }
+}